requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pinecone-client
openai
anthropic
httpx
